requirements.txt
adal==1.2.7
aiofiles==0.8.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
azure-cognitiveservices-search-websearch==2.0.0
azure-common==1.1.28
azure-core==1.29.7
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
discord.py==2.3.2
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isodate==0.6.1
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
oauthlib==3.2.2
openai==1.8.0
outcome==1.3.0.post0
praw==7.7.1
prawcore==2.4.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
update-checker==0.18.0
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
